refactor: simplify client screens
