+++ packages/apps/crm/src/components/organisms/Client/ClientBottom/ClientBottom.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import {EditButton} from '../../../organisms';
+
+const ClientBottom = ({navigation, idClient}) => {
+  return (
+    <EditButton
+      onPress={() =>
+        navigation.navigate('ClientFormScreen', {
+          idClient: idClient,
+        })
+      }
+    />
+  );
+};
+
+export default ClientBottom;
