+++ packages/apps/crm/src/components/organisms/Client/index.js
@@ -1,0 +1,3 @@
+export {default as ClientBottom} from './ClientBottom/ClientBottom';
+export {default as ClientDropdownCards} from './ClientDropdownCards/ClientDropdownCards';
+export {default as ClientHeader} from './ClientHeader/ClientHeader';
