+++ packages/apps/crm/src/screens/client/ClientDetailsScreen.js
@@ -1,19 +1,24 @@
 import React, {useEffect} from 'react';
 import {ScrollView} from 'react-native';
-import {Screen, HeaderContainer} from '@axelor/aos-mobile-ui';
+import {Screen, HeaderContainer, NotesCard} from '@axelor/aos-mobile-ui';
 import {
   useSelector,
   HeaderOptionsMenu,
   useDispatch,
+  useTranslator,
 } from '@axelor/aos-mobile-core';
-import {ClientBody, ClientBottom, ClientHeader} from '../../components';
-import {searchContactById} from '../../features/contactSlice';
-import {fetchPartnerEventById} from '../../features/eventSlice';
+import {
+  ClientBottom,
+  ClientDropdownCards,
+  ClientHeader,
+} from '../../components';
 import {getClientbyId} from '../../features/clientSlice';
 
 const ClientDetailsScreen = ({navigation, route}) => {
   const idClient = route.params.idClient;
+  const I18n = useTranslator();
   const dispatch = useDispatch();
+
   const {mobileSettings} = useSelector(state => state.config);
   const {client} = useSelector(state => state.client);
 
@@ -35,20 +40,12 @@ const ClientDetailsScreen = ({navigation, route}) => {
     dispatch(getClientbyId({clientId: idClient}));
   }, [dispatch, idClient]);
 
-  useEffect(() => {
-    const idList = client.contactPartnerSet?.map(item => item.id);
-    dispatch(searchContactById(idList));
-  }, [dispatch, client.contactPartnerSet]);
-
-  useEffect(() => {
-    dispatch(fetchPartnerEventById(idClient));
-  }, [dispatch, idClient]);
-
   return (
     <Screen removeSpaceOnTop={true}>
       <HeaderContainer expandableFilter={false} fixedItems={<ClientHeader />} />
       <ScrollView>
-        <ClientBody navigation={navigation} />
+        <NotesCard title={I18n.t('Crm_Notes')} data={client.description} />
+        <ClientDropdownCards navigation={navigation} />
       </ScrollView>
       <ClientBottom idClient={idClient} navigation={navigation} />
     </Screen>
